requirements.txt
discord.py
python-dotenv
aiohttp
requests
configparser
flask